Make Types::Struct class that mixes in Composable and Attributes
