+++ lib/plumb.rb
@@ -76,6 +76,7 @@ require 'plumb/array_class'
 require 'plumb/stream_class'
 require 'plumb/hash_class'
 require 'plumb/interface_class'
+require 'plumb/attributes'
 require 'plumb/types'
 require 'plumb/json_schema_visitor'
 require 'plumb/schema'
